Create post-processing function
